pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iac-audit-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.60",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "python-hcl2>=4.3.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
iac-audit = "iac_audit_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/iac_audit_agent"]
